refactor: rename TpoProfile to TreeCounter and update imports across components
