+++ src/tenants/salesforce/VTOCampaign2025/components/Landing.tsx
@@ -1,5 +1,5 @@
 import styles from './../styles/Landing.module.scss';
-import TreeCounter from '../../TreeCounter/TreeCounter';
+import TreeCounter from '../../TreeCounter';
 
 interface Props {
   tenantScore?: { total: number };
